wip android

+++ src/Plugins/DevelopmentConverterPlugin/HotspotImageConverterPNGToHIT.cpp
@@ -11,7 +11,7 @@
 
 #include "Kernel/Logger.h"
 
-#include <math.h>
+#include <cmath>
 #include <algorithm>
 
 namespace Mengine
@@ -200,8 +200,8 @@ namespace Mengine
     //////////////////////////////////////////////////////////////////////////
     uint32_t HotspotImageConverterPNGToHIT::calcMimMapBufferLevelAndSize_( uint32_t _width, uint32_t _height, uint32_t & _level )
     {
-        uint32_t mipmap_pow_width = (uint32_t)((std::log)((float)_width) / (std::log)(2.f));
-        uint32_t mipmap_pow_height = (uint32_t)((std::log)((float)_height) / (std::log)(2.f));
+        uint32_t mipmap_pow_width = (uint32_t)((std::logf)((float)_width) / (std::logf)(2.f));
+        uint32_t mipmap_pow_height = (uint32_t)((std::logf)((float)_height) / (std::logf)(2.f));
 
         uint32_t mipmap_level = (std::min)(mipmap_pow_width, mipmap_pow_height) + 1;
